Api Reto levantada
ApiRetolevantada

+++ src/main/java/retodaw/dtos/SolicitudDto.java
@@ -14,7 +14,7 @@ public class SolicitudDto {
     private Date fecha;
     private String archivo;
     private String comentarios;
-    private int estado;
+    private Byte estado;
     private String curriculum;
     private String email;
     private int idVacante;
